% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrLoadings.R
\name{makeCorrLoadings}
\alias{makeCorrLoadings}
\title{Generate a correlation matrix of inter-item correlations from
item factor loadings}
\usage{
makeCorrLoadings(
  loadings,
  factorCor = NULL,
  uniquenesses = NULL,
  nearPD = FALSE
)
}
\arguments{
\item{loadings}{(numeric matrix) \strong{k} (items) by \strong{f} (factors)
matrix of \strong{standardised} factor loadings. Item names and Factor names
can be taken from the row_names (items) and the column_names (factors),
if present.}

\item{factorCor}{(numeric matrix) \strong{f x f} factor correlation matrix}

\item{uniquenesses}{(numeric vector) length \strong{k} vector of uniquenesses.
If NULL, the default, compute from the calculated communalities.}

\item{nearPD}{(logical) If TRUE, project factorCor and the final
correlation matrix onto nearest Positive Definite matrix, if needed.
(It should never be needed.)}
}
\value{
Correlation matrix of inter-item correlations
}
\description{
\code{makeCorrLoadings()} generates a correlation matrix of
inter-item correlations based on item factor loadings as might be seen in
\emph{Exploratory Factor Analysis} (\strong{EFA}) or a \emph{Structural Equation Model}
(\strong{SEM}).

Such a correlation matrix can be applied to the \code{makeItems()}
function to generate synthetic data with those predefined factor structures.
}
\examples{

# Example loadings
factorLoadings <- matrix(
  c(
      0.05, 0.20, 0.70,
      0.10, 0.05, 0.80,
      0.05, 0.15, 0.85,
      0.20, 0.85, 0.15,
      0.05, 0.85, 0.10,
      0.10, 0.90, 0.05,
      0.90, 0.15, 0.05,
      0.80, 0.10, 0.10
   ),
   nrow = 8, ncol = 3, byrow = TRUE
)

 # row and column names
 rownames(factorLoadings) <- c("Q1", "Q2", "Q3", "Q4", "Q5", "Q6", "Q7", "Q8")
 colnames(factorLoadings) <- c("Factor1", "Factor2", "Factor3")

 # Factor correlation matrix
 factorCor <- matrix(
 c(
   1.0,  0.7, 0.6,
   0.7,  1.0, 0.4,
   0.6,  0.4, 1.0
  ),
 nrow = 3, byrow = TRUE
 )

 # Apply the function
 itemCorrelations <- makeCorrLoadings(factorLoadings, factorCor)

 round(itemCorrelations, 3)


}
