% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItemsScale.R
\name{makeItemsScale}
\alias{makeItemsScale}
\title{Generate scale items from a summated scale, with desired Cronbach's Alpha}
\usage{
makeItemsScale(
  scale,
  lowerbound,
  upperbound,
  items,
  alpha = 0.8,
  summated = TRUE,
  progress = TRUE
)
}
\arguments{
\item{scale}{(int) a vector or dataframe of the summated rating scale.
Should range from \eqn{lowerbound \times items} to
\eqn{upperbound \times items}}

\item{lowerbound}{(int) lower bound of the scale item
(example: '1' in a '1' to '5' rating)}

\item{upperbound}{(int) upper bound of the scale item
(example: '5' in a '1' to '5' rating)}

\item{items}{(positive, int) k, or number of columns to generate}

\item{alpha}{(posiitve, real) desired \emph{Cronbach's Alpha} for the
new dataframe of items.
Default = '0.8'.

See \verb{@details} for further information on the \code{alpha} parameter}

\item{summated}{logical. If TRUE (default), \code{scale} is assumed to be a
summated scale. If FALSE, the input is treated as a mean scale and will
be multiplied by \code{items} internally.}

\item{progress}{logical. If TRUE (default is \code{interactive()}), a progress
bar is displayed during optimisation.
Set to FALSE for slightly faster execution.}
}
\value{
a dataframe with 'items' columns and 'length(scale)' rows
}
\description{
\code{makeItemsScale()} generates a random dataframe
of scale items based on a predefined summated scale
(such as created by the \code{\link[=lfast]{lfast()}} function),
and a desired \emph{Cronbach's Alpha}.
}
\details{
\subsection{alpha}{

\code{makeItemsScale()} takes each value of a vector of Likert scales and
produces a row of 'k' values that average the given scale value, and
then rearranges the item values within each row,
attempting to give a dataframe of Likert-scale items that produce a
predefined \emph{Cronbach's Alpha}.

Default value for target alpha is '0.8'.

More extreme values for the 'variance' parameter may reduce the chances
of achieving the desired Alpha. So you may need to experiment a little.

Candidate item combinations are sampled based on their dispersion,
favouring moderate variability. Final item properties emerge through
optimisation to satisfy reliability constraints.
}
}
\examples{

summatedScale <- c(16, 11, 14, 7, 10, 13, 16, 13, 12, 14, 8, 11, 8, 18,  8, 13)

newItems <- makeItemsScale(
  scale = summatedScale,
  lowerbound = 1, upperbound = 5,
  items = 4
)


}
