% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{print.likert_reliability}
\alias{print.likert_reliability}
\title{Print method for reliability objects}
\usage{
\method{print}{likert_reliability}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=reliability]{reliability()}}.}

\item{...}{Unused.}
}
\description{
Print method for reliability objects
}
\keyword{internal}
