% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScalesRegression.R
\name{print.makeScalesRegression}
\alias{print.makeScalesRegression}
\title{Print method for makeScalesRegression objects}
\usage{
\method{print}{makeScalesRegression}(x, ...)
}
\arguments{
\item{x}{An object of class "makeScalesRegression"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for makeScalesRegression objects
}
