% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexact.R
\name{lexact}
\alias{lexact}
\title{Generate rating-scale data with only Mean and Standard Deviation}
\arguments{
\item{n}{the number of observations to simulate}

\item{mean}{target mean}

\item{sd}{target standard deviation}

\item{lowerbound}{a lower bound for the data to be generated}

\item{upperbound}{an upper bound for the data to be generated}

\item{items}{number of items in the Likert scale. Default = 1}

\item{seed}{optional seed for reproducibility}
}
\value{
a vector of simulated data with user-specified conditions.
}
\description{
\code{lexact()} generates rating-scale values with
predefined first and second moments.
}
\details{
If feasible, moments are exact to two decimal places.
}
\examples{

x <- lexact(
  n = 16,
  mean = 3.2,
  sd = 0.85,
  lowerbound = 1,
  upperbound = 5,
  items = 6
)

x <- lexact(
  n = 16,
  mean = 1.2,
  sd = 1.00,
  lowerbound = -3,
  upperbound = 3,
  items = 4
)

x <- lexact(16, 2, 2.5, 0, 10)

}
