% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexact.R
\name{lexact}
\alias{lexact}
\title{Generate Likert scale data with only Mean and Standard Deviation}
\arguments{
\item{n}{the number of observations to simulate}

\item{mean}{target mean}

\item{sd}{target standard deviation}

\item{lowerbound}{a lower bound for the data to be generated}

\item{upperbound}{an upper bound for the data to be generated}

\item{items}{number of items in the Likert scale. Default = 1}

\item{seed}{optional seed for reproducibility}
}
\value{
a vector of simulated data with user-specified conditions.
}
\description{
\code{lexact()} is a function to randomly generate rating-scale values with predefined first and second moments.
}
\details{
If feasible, the function should produce data with moments that are exact to two decimal places.
}
\examples{

x <- lexact(n = 32, mean = 4.2, sd = 0.85, lowerbound = 1, upperbound = 7, items = 6)
x <- lexact(32, 4, 1.5, 0, 10, 1)
x <- lexact(16, 3.5, 1.0, 1, 5, 5)
x <- lexact(n = 64, mean = 1.2, sd = 1.00, lowerbound = -3, upperbound = 3, items = 6)

}
