% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexact.R
\name{lexact}
\alias{lexact}
\title{Deprecated. Use lfast() instead}
\usage{
lexact(n, mean, sd, lowerbound, upperbound, items = 1)
}
\arguments{
\item{n}{(positive, int) number of observations to generate}

\item{mean}{(real) target mean}

\item{sd}{(real) target standard deviation}

\item{lowerbound}{(positive, int) lower bound}

\item{upperbound}{(positive, int) upper bound}

\item{items}{(positive, int) number of items in the rating scale.}
}
\value{
a vector of simulated data approximating user-specified conditions.
}
\description{
\code{lexact} is DEPRECATED.
Replaced in LikertMakeR Version 0.4.0 by new version of \code{lfast}.

\code{lexact} remains as a legacy for earlier package users.
It is now just a wrapper for \code{lfast}

Previously, \code{lexact} used a Differential Evolution (DE) algorithm to
find an optimum solution with desired mean and standard deviation,
but we found that the updated \code{lfast} function is much faster and just
as accurate.

Also the package is much less bulky.
}
\examples{

x <- lexact(
  n = 256,
  mean = 4.0,
  sd = 1.0,
  lowerbound = 1,
  upperbound = 7,
  items = 6
)

x <- lexact(256, 2, 1.8, 0, 10)

}
