% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfast.R
\name{lfast}
\alias{lfast}
\title{Rating scale data (e.g. Likert scale) from a Scaled Beta Distribution}
\usage{
lfast(n, mean, sd, lowerbound, upperbound, items = 1, seed)
}
\arguments{
\item{n}{number of observations to generate}

\item{mean}{target mean}

\item{sd}{target standard deviation}

\item{lowerbound}{lower bound (e.g. '1' for a 1-5 rating scale)}

\item{upperbound}{upper bound (e.g. '5' for a 1-5 rating scale)}

\item{items}{number of items in the rating scale. Default = 1}

\item{seed}{optional seed for reproducibility}
}
\value{
a vector of simulated data approximating user-specified conditions.
}
\description{
\code{lfast()} generates random discrete values from a
(scaled Beta distribution) so the data replicate a rating scale -
for example,a 1-5 scale made from 5 items (questions) or 0-10
likelihood-of-purchase scale.
}
\examples{

x <- lfast(
  n = 256,
  mean = 4.0,
  sd = 1.0,
  lowerbound = 1,
  upperbound = 7,
  items = 6
)

x <- lfast(256, 2, 1.8, 0, 10)

x <- lfast(256, 2, 1.0, 1, 5, 5)

}
