% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItemsScale.R
\name{makeItemsScale}
\alias{makeItemsScale}
\title{scale items from a summated scale}
\usage{
makeItemsScale(scale, lowerbound, upperbound, items)
}
\arguments{
\item{scale}{(int) a vector or dataframe of the summated rating scale.
Should range from ('lowerbound' * 'items') to ('upperbound' * 'items')}

\item{lowerbound}{(int) lower bound of the scale item
(example: '1' in a '1' to '5' rating)}

\item{upperbound}{(int) upper bound of the scale item
(example: '5' in a '1' to '5' rating)}

\item{items}{(positive, int) k, or number of columns to generate}
}
\value{
a dataframe with 'items' columns and 'length(scale)' rows
}
\description{
\code{makeItemsScale()} generates a random dataframe
of scale items based on a predefined summated scale,
such as created by the \code{lfast()} function.

scale, lowerbound, upperbound, items
}
\examples{

# define parameters
items <- 4
lowerbound <- 1
upperbound <- 5

## scale properties
n <- 64
mean <- 3.5
sd <- 1.00

## create scale
set.seed(42)
meanScale <- lfast(n = n, mean = mean, sd = sd,
     lowerbound = lowerbound, upperbound = upperbound,
     items = items)
summatedScale <- meanScale * items

## create items
newItems <- makeItemsScale(scale = summatedScale,
     lowerbound = lowerbound, upperbound = upperbound,
     items = items)
head(newItems)



lowerbound <- 1
upperbound <- 7
items <- 3

myvalues <- c((lowerbound * items):(upperbound * items))

mydata <- makeItemsScale(scale = myvalues,
     lowerbound = lowerbound, upperbound = upperbound,
     items = items)

mydata

}
